Modal Confirmation Box and RecipeResolver bug fix
Created a modal confirmation box for confirmation of whether to save data in the backend or not. This was done programmatically.

Small bug fix where having no recipes in the backend caused an error in the reciperesolver service. This was fixed by making the first recipe undeletable ergo the backend recipe list will never be empty.